requirements/base.txt
# Web Scraping
beautifulsoup4==4.12.0
requests==2.31.0
selenium==4.15.2

# Data Processing
pandas==2.1.1
numpy==1.24.3
nltk==3.8.1
scikit-learn==1.3.0

# Database
pymongo==4.5.0
motor==3.3.1
psycopg2-binary==2.9.10

# API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2

# AWS
boto3==1.28.44
aws-lambda-powertools==2.26.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2

# AI/ML
pinecone-client>=3.0.0
openai>=1.0.0
langchain>=0.1.0
tiktoken>=0.5.0 